requirements.txt
requests
pandas
pandas-ta
xgboost
lightgbm
scikit-learn
optuna
matplotlib
flask
python-dotenv
